pyproject.toml
[tool.poetry]
name = "piscuss"
version = "0.1.0"
description = "Piscuss is a fast and lightweight open source discussion application."
authors = ["Ethan V"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
Django = "^4.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
